css/css-fonts/parsing/font-variant-emoji-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-variant-emoji with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#propdef-font-variant-emoji">
<meta name="assert" content="font-variant-emoji supports only the grammar 'auto | text | emoji | unicode'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-variant-emoji', 'normal');
test_invalid_value('font-variant-emoji', 'auto text');
test_invalid_value('font-variant-emoji', 'emoji, unicode');
</script>
</body>
</html>
